+++ egcg_core/rest_communication.py
@@ -1,6 +1,6 @@
+import json
 import mimetypes
 import requests
-import json
 from urllib.parse import urljoin
 from egcg_core.config import cfg
 from egcg_core.app_logging import AppLogger
@@ -146,10 +146,12 @@ class Communicator(AppLogger):
         content = self.get_content(endpoint, paginate, quiet, **query_args)
         elements = content['data']
 
-        if all_pages and 'next' in content['_links']:
-            next_query = self._parse_query_string(content['_links']['next']['href'], requires=('max_results', 'page'))
-            query_args.update(next_query)
-            elements.extend(self.get_documents(endpoint, all_pages=True, quiet=quiet, **query_args))
+        if all_pages:
+            while 'next' in content['_links']:
+                next_query = self._parse_query_string(content['_links']['next']['href'], requires=('max_results', 'page'))
+                query_args.update(next_query)
+                content = self.get_content(endpoint, paginate, quiet, **query_args)
+                elements.extend(content['data'])
 
         return elements
 
@@ -160,9 +162,12 @@ class Communicator(AppLogger):
         else:
             self.warning('No document found in endpoint %s for %s', endpoint, query_args)
 
-    def post_entry(self, endpoint, payload):
+    def post_entry(self, endpoint, payload, use_data=False):
         files, payload = self._detect_files_in_json(payload)
-        return self._req('POST', self.api_url(endpoint), json=payload, files=files)
+        if use_data:
+            return self._req('POST', self.api_url(endpoint), data=payload, files=files)
+        else:
+            return self._req('POST', self.api_url(endpoint), json=payload, files=files)
 
     def put_entry(self, endpoint, element_id, payload):
         files, payload = self._detect_files_in_json(payload)